requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.3.5
idna==3.3
Markdown==3.3.6
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
django-webpack-loader==1.5.0
